Remove Product + TokenHelper

+++ src/main/java/fr/miage/samba/backend/web/controller/ProductController.java
@@ -1,10 +1,16 @@
 package fr.miage.samba.backend.web.controller;
 
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.algorithms.Algorithm;
+import com.auth0.jwt.interfaces.DecodedJWT;
 import com.fasterxml.jackson.databind.ser.FilterProvider;
 import com.fasterxml.jackson.databind.ser.impl.SimpleBeanPropertyFilter;
 import com.fasterxml.jackson.databind.ser.impl.SimpleFilterProvider;
 import fr.miage.samba.backend.model.ProductDto;
+import fr.miage.samba.backend.model.UserDto;
+import fr.miage.samba.backend.security.TokenHelper;
 import fr.miage.samba.backend.services.ProductService;
+import fr.miage.samba.backend.services.UserService;
 import fr.miage.samba.backend.web.exceptions.EmptyField;
 import fr.miage.samba.backend.web.exceptions.IncorrectProductPrice;
 import fr.miage.samba.backend.web.exceptions.ProductNotFound;
@@ -12,6 +18,7 @@ import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.json.MappingJacksonValue;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
@@ -22,6 +29,8 @@ import java.util.List;
 import java.util.Optional;
 
 import static fr.miage.samba.backend.security.SecurityConstants.HEADER_STRING;
+import static fr.miage.samba.backend.security.SecurityConstants.SECRET;
+import static fr.miage.samba.backend.security.SecurityConstants.TOKEN_PREFIX;
 
 @RequestMapping(value ="/products")
 @RestController
@@ -29,6 +38,9 @@ public class ProductController {
 
     @Autowired
     private ProductService productService;
+
+    @Autowired
+    private UserService userService;
 
     //Get all
     @GetMapping()
@@ -45,13 +57,29 @@ public class ProductController {
             throw new ProductNotFound();
         }
 
-        return productService.getDetailsOfProduitById(id).get();
+        return requestResult.get();
     }
 
     //Get Specified
     @DeleteMapping("/{id}")
-    public void removeProductById(@PathVariable String id){
-        this.productService.remove(id);
+    public void removeProductById(HttpServletRequest request,@PathVariable String id){
+
+        Optional<ProductDto> requestResult = productService.getDetailsOfProduitById(id);
+
+        if(!requestResult.isPresent()){
+            throw new ProductNotFound();
+        }
+
+        String userId = TokenHelper.extractSubjectOf(request);
+        String sellerId = requestResult.get().getSellerId();
+        UserDto user = userService.getUserByUsername(userId);
+
+        if(sellerId != null && !sellerId.trim().isEmpty() && user != null
+                && requestResult.get().getSellerId().equals(user.getId())){
+            this.productService.remove(id);
+        }else{
+            throw new AccessDeniedException("You aren't allow to delete this element");
+        }
     }
 
     //Ajouter un produit
@@ -61,7 +89,7 @@ public class ProductController {
         if(product.getDescription().isEmpty() || product.getTitle().isEmpty() || product.getSellerId().isEmpty()){
             throw new EmptyField();
         }
-        if(product.getPrix() <= 0){
+        if(product.getPrice() <= 0){
             throw new IncorrectProductPrice();
         }
         product.setId(ObjectId.get());
@@ -75,7 +103,6 @@ public class ProductController {
                 .path("/{id}")
                 .buildAndExpand(productAdded.getId())
                 .toUri();
-
 
         return ResponseEntity.created(location).body(product);
     }
